refactor: レビュー指摘に対応（レスポンス変換の配置・エラーテストの集約）
- article-response.ts は handler ではなくレスポンス変換のため handler/ 直下から
  server/article/ 直下へ移動
- エラー系のみで分割していたテストファイルを廃止し、各ハンドラーの
  テストファイルへ集約。正常系は実装へ委譲し spy で当該メソッドのみ差し替える
